backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.17.0
pydantic-settings==2.10.1
google-genai==1.36.0
google-api-python-client==2.194.0
google-auth==2.49.2
python-multipart==0.0.20
